Fix props spreading
