+++ src/app/app.scss
@@ -41,7 +41,7 @@
   .ant-layout {
     &.App {
       .layout-content-article {
-        background-color: #1f1f1f;
+        background-color: $dark-main;
       }
     }
   }
